Separated GUI Frames into their own file to mitigate search time Added is_char_{printable, valid} into GUI method (instead of standalone function)
